fix(cliente): altera tipo do atributo cpf
- muda tipo do cpf de int para long (suporta 11 digitos, tamanho do cpf)

+++ nexthome/src/main/java/com/fatec/nexthome/Cliente.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @Entity
 public class Cliente {
     @Id
-    private int cpf;
+    private long cpf;
     private String nome;
     private String email;
     private String senha;
